requirements.txt
absl-py==0.9.0
alabaster==0.7.12
albumentations==0.4.5
alembic==1.4.2
apache-airflow==1.10.10
apispec==1.3.3
argcomplete==1.11.1
astor==0.8.1
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
bleach==3.1.5
cached-property==1.5.1
cattrs==0.9.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.32
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
dnspython==1.16.0
docutils==0.16
email-validator==1.1.1
entrypoints==0.3
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.3
flask-swagger==0.2.13
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.18.2
gast==0.3.3
graphviz==0.14
grpcio==1.29.0
gunicorn==19.10.0
h5py==2.10.0
idna==2.9
imageio==2.8.0
imagesize==1.2.0
imgaug==0.2.6
importlib-metadata==1.6.0
imutils==0.5.3
ipykernel==5.3.0
ipyparallel==6.3.0
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso8601==0.1.12
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.10.3
joblib==0.15.1
json-merge-patch==0.2
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.1.3
jupyterlab-server==1.1.5
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.3
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
matplotlib==3.2.1
mistune==0.8.4
natsort==7.0.1
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
nose==1.3.7
notebook==6.0.3
numpy==1.16.5
opencv-contrib-python==4.2.0.34
opencv-python-headless==4.2.0.34
packaging==20.4
pandas==0.25.3
pandocfilters==1.4.2
parso==0.7.0
pendulum==1.4.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prison==0.1.3
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.1
psutil==5.7.0
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-daemon==2.1.2
python-dateutil==2.8.1
python-editor==1.0.4
python3-openid==3.1.0
pytz==2020.1
pytzdata==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
requests==2.23.0
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
Send2Trash==1.5.0
setproctitle==1.1.10
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.0.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.17
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.6
tabulate==0.8.7
tenacity==4.12.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.2
threadpoolctl==2.1.0
thrift==0.13.0
tifffile==2020.5.25
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
typing==3.7.4.1
typing-extensions==3.7.4.2
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
WTForms==2.3.1
zipp==3.1.0
zope.deprecation==4.4.0
